remove tetex dependency
tetex is not required by PyX. PyX works fine as a high quality
scriptable pdf generator even in server contexts where the tetex
dependency is a pain.

tex (i.e. parts of tetex) is needed only for typesetting text with PyX,
and PyX users are very well aware of that recommended dependency.

So, remove the strict one.

see bz #760524